feat(utils): move getInitials function to utils and update imports
