+++ src/components/common/index.tsx
@@ -14,4 +14,4 @@ export {
 export { default as DevUserSwitcher } from "./dev-user-switcher";
 export { default as Footer } from "./footer";
 export { default as Header } from "./header";
-export { getInitials, default as UserMenu } from "./user-menu";
+export { default as UserMenu } from "./user-menu";
