+++ src/stories/user-menu.stories.tsx
@@ -2,7 +2,6 @@ import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 import { ChevronDown, LogOut, User, Wallet } from "lucide-react";
 import Link from "next/link";
 
-import { getInitials } from "@/components/common";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
@@ -12,6 +11,7 @@ import {
   HoverCardTrigger,
 } from "@/components/ui/hover-card";
 import { Separator } from "@/components/ui/separator";
+import { getInitials } from "@/lib/utils/string";
 
 // Mock component for Storybook that accepts controlled props
 interface UserMenuProps {
